Simplify Package API
