+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/PackageVersionApi.java
@@ -18,6 +18,9 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.core.data.Constants;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -29,15 +32,21 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/mvn/package_version")
 public class PackageVersionApi {
 
-    private final PackageVersionApiService service;
-
-    public PackageVersionApi(PackageVersionApiService service) {
-        this.service = service;
-    }
 
     @GetMapping(value = "/{id}/rcg", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getERCGLink(@PathVariable("id") long packageVersionId) {
-        return service.getERCGLink(packageVersionId);
+        var artifact = KnowledgeBaseConnector.kbDao.getArtifactName(packageVersionId);
+        if (artifact == null) {
+            return new ResponseEntity<>("Package version not found", HttpStatus.NOT_FOUND);
+        }
+        var coordinate = artifact.split(Constants.mvnCoordinateSeparator);
+        var groupId = coordinate[0];
+        var artifactId = coordinate[1];
+        var version = coordinate[2];
+        String result = String.format("%smvn/%s/%s/%s_%s_%s.json", KnowledgeBaseConnector.rcgBaseUrl,
+                artifactId.charAt(0), artifactId, artifactId, groupId, version);
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
