+++ analyzer/restapi-plugin/src/test/java/eu/fasten/analyzer/restapiplugin/mvn/api/PackageApiTest.java
@@ -18,8 +18,12 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
 import eu.fasten.analyzer.restapiplugin.mvn.RestApplication;
+import eu.fasten.core.data.metadatadb.MetadataDao;
+import eu.fasten.core.maven.data.PackageVersionNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.http.HttpStatus;
@@ -28,100 +32,189 @@ import org.springframework.http.ResponseEntity;
 import java.io.IOException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-class PackageApiTest {
+public class PackageApiTest {
 
-    private PackageApiService service;
-    private PackageApi api;
+    private PackageApi service;
+    private MetadataDao kbDao;
     private final int offset = 0;
     private final int limit = Integer.parseInt(RestApplication.DEFAULT_PAGE_SIZE);
 
     @BeforeEach
     void setUp() {
-        service = Mockito.mock(PackageApiService.class);
-        api = new PackageApi(service);
+        service = new PackageApi();
+        kbDao = Mockito.mock(MetadataDao.class);
+        KnowledgeBaseConnector.kbDao = kbDao;
     }
 
     @Test
     void getAllPackagesTest() {
-        var response = new ResponseEntity<>("all packages", HttpStatus.OK);
-        Mockito.when(service.getAllPackages(offset, limit)).thenReturn(response);
-        var result = api.getAllPackages(offset, limit);
-        assertEquals(response, result);
-        Mockito.verify(service).getAllPackages(offset, limit);
+        var response = "all packages";
+        Mockito.when(kbDao.getAllPackages(offset, limit)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getAllPackages(offset, limit);
+        assertEquals(expected, result);
+        Mockito.verify(kbDao).getAllPackages(offset, limit);
     }
 
     @Test
-    void getPackageLastVersionTest() {
-        var packageName = "pkg name";
-        var response = new ResponseEntity<>("latest package version", HttpStatus.OK);
-        Mockito.when(service.getPackageLastVersion(packageName)).thenReturn(response);
-        var result = api.getPackageLastVersion(packageName);
-        assertEquals(response, result);
-        Mockito.verify(service).getPackageLastVersion(packageName);
+    void getPackageLastVersionPositiveTest() {
+        var packageName = "group:artifact";
+        var response = "a package";
+        Mockito.when(kbDao.getPackageLastVersion(packageName)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getPackageLastVersion(packageName);
+        assertEquals(expected, result);
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageLastVersion(packageName);
+    }
+
+    @Test
+    void getPackageLastVersionNegativeTest() {
+        var packageName = "group:artifact";
+        Mockito.when(kbDao.getPackageLastVersion(packageName)).thenReturn(null);
+        var result = service.getPackageLastVersion(packageName);
+        assertEquals(HttpStatus.NOT_FOUND, result.getStatusCode());
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageLastVersion(packageName);
     }
 
     @Test
     void getPackageVersionsTest() {
-        var packageName = "pkg name";
-        var response = new ResponseEntity<>("packages", HttpStatus.OK);
-        Mockito.when(service.getPackageVersions(packageName, offset, limit)).thenReturn(response);
-        var result = api.getPackageVersions(packageName, offset, limit);
-        assertEquals(response, result);
-        Mockito.verify(service).getPackageVersions(packageName, offset, limit);
+        var packageName = "group:artifact";
+        var response = "package versions";
+        Mockito.when(kbDao.getPackageVersions(packageName, offset, limit)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getPackageVersions(packageName, offset, limit);
+        assertEquals(expected, result);
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageVersions(packageName, offset, limit);
     }
 
     @Test
-    void getPackageVersionTest() throws IOException {
-        var packageName = "pkg name";
-        var version = "pkg version";
-        var response = new ResponseEntity<>("package version", HttpStatus.OK);
-        Mockito.when(service.getPackageVersion(packageName, version, null, null)).thenReturn(response);
-        var result = api.getPackageVersion(packageName, version, null, null);
-        assertEquals(response, result);
-        Mockito.verify(service).getPackageVersion(packageName, version, null, null);
+    void getPackageVersionPositiveTest() {
+        var packageName = "group:artifact";
+        var version = "version";
+        var response = "package version";
+        Mockito.when(kbDao.getPackageVersion(packageName, version)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getPackageVersion(packageName, version, null, null);
+        assertEquals(expected, result);
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageVersion(packageName, version);
     }
 
     @Test
-    void getPackageMetadataTest() {
-        var packageName = "pkg name";
-        var version = "pkg version";
-        var response = new ResponseEntity<>("package version metadata", HttpStatus.OK);
-        Mockito.when(service.getPackageMetadata(packageName, version)).thenReturn(response);
-        var result = api.getPackageMetadata(packageName, version);
-        assertEquals(response, result);
-        Mockito.verify(service).getPackageMetadata(packageName, version);
+    void getPackageVersionNegativeTest() {
+        var packageName = "group:artifact";
+        var version = "version";
+        Mockito.when(kbDao.getPackageVersion(packageName, version)).thenReturn(null);
+        var result = service.getPackageVersion(packageName, version, null, null);
+        assertEquals(HttpStatus.BAD_REQUEST, result.getStatusCode());
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageVersion(packageName, version);
     }
 
     @Test
-    void getPackageCallgraphTest() throws IOException {
-        var packageName = "pkg name";
-        var version = "pkg version";
-        var response = new ResponseEntity<>("package version callgraph", HttpStatus.OK);
-        Mockito.when(service.getPackageCallgraph(packageName, version, offset, limit, null, null)).thenReturn(response);
-        var result = api.getPackageCallgraph(packageName, version, offset, limit, null, null);
-        assertEquals(response, result);
-        Mockito.verify(service).getPackageCallgraph(packageName, version, offset, limit, null, null);
+    void getPackageVersionIngestionTest() {
+        var packageName = "junit:junit";
+        var version = "4.12";
+        Mockito.when(kbDao.getPackageVersion(packageName, version)).thenReturn(null);
+        var result = service.getPackageVersion(packageName, version, null, null);
+        assertEquals(HttpStatus.CREATED, result.getStatusCode());
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageVersion(packageName, version);
     }
 
     @Test
-    void searchPackageNamesTest() {
-        var packageName = "pkg name";
-        var response = new ResponseEntity<>("matching packages", HttpStatus.OK);
-        Mockito.when(service.searchPackageNames(packageName, offset, limit)).thenReturn(response);
-        var result = api.searchPackageNames(packageName, offset, limit);
-        assertEquals(response, result);
-        Mockito.verify(service).searchPackageNames(packageName, offset, limit);
+    void getPackageMetadataPositiveTest() {
+        var packageName = "group:artifact";
+        var version = "version";
+        var response = "package metadata";
+        Mockito.when(kbDao.getPackageMetadata(packageName, version)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getPackageMetadata(packageName, version);
+        assertEquals(expected, result);
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageMetadata(packageName, version);
     }
 
     @Test
-    void getERCGLinkTest() throws IOException {
-        var packageName = "pkg name";
-        var version = "pkg version";
-        var response = new ResponseEntity<>("link to ercg", HttpStatus.OK);
-        Mockito.when(service.getERCGLink(packageName, version, null, null)).thenReturn(response);
-        var result = api.getERCGLink(packageName, version, null, null);
-        assertEquals(response, result);
-        Mockito.verify(service).getERCGLink(packageName, version, null, null);
+    void getPackageNegativeMetadataTest() {
+        var packageName = "group:artifact";
+        var version = "version";
+        Mockito.when(kbDao.getPackageMetadata(packageName, version)).thenReturn(null);
+        var result = service.getPackageMetadata(packageName, version);
+        assertEquals(HttpStatus.NOT_FOUND, result.getStatusCode());
+
+        Mockito.verify(kbDao, Mockito.times(1)).getPackageMetadata(packageName, version);
+    }
+
+    @Test
+    void getPackageCallgraphPositiveTest() throws IOException {
+        var packageName = "group:artifact";
+        var version = "version";
+        var response = "package callgraph";
+        Mockito.when(kbDao.getPackageCallgraph(packageName, version, offset, limit)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.getPackageCallgraph(packageName, version, offset, limit, null, null);
+        Mockito.verify(kbDao).getPackageCallgraph(packageName, version, offset, limit);
+        assertEquals(expected, result);
+    }
+
+    @Test
+    void getPackageCallgraphIngestionTest() throws IOException {
+        var packageName = "junit:junit";
+        var version = "4.12";
+        Mockito.when(kbDao.getPackageCallgraph(packageName, version, offset, limit)).thenThrow(new PackageVersionNotFoundException("Error"));
+        var result = service.getPackageCallgraph(packageName, version, offset, limit, null, null);
+        assertEquals(HttpStatus.CREATED, result.getStatusCode());
+
+        Mockito.verify(kbDao).getPackageCallgraph(packageName, version, offset, limit);
+    }
+
+    @Test
+    void searchPackageNamesPositiveTest() {
+        var packageName = "group:artifact";
+        var response = "matching package versions";
+        Mockito.when(kbDao.searchPackageNames(packageName, offset, limit)).thenReturn(response);
+        var expected = new ResponseEntity<>(response, HttpStatus.OK);
+        var result = service.searchPackageNames(packageName, offset, limit);
+        assertEquals(expected, result);
+        Mockito.verify(kbDao).searchPackageNames(packageName, offset, limit);
+    }
+
+    @Test
+    void searchPackageNamesNegativeTest() {
+        var packageName = "group:artifact";
+        Mockito.when(kbDao.searchPackageNames(packageName, offset, limit)).thenReturn(null);
+        var result = service.searchPackageNames(packageName, offset, limit);
+        assertEquals(HttpStatus.NOT_FOUND, result.getStatusCode());
+
+        Mockito.verify(kbDao).searchPackageNames(packageName, offset, limit);
+    }
+
+    @Test
+    void getERCGLinkPositiveTest() throws IOException {
+        var packageName = "junit:junit";
+        var version = "4.12";
+        Mockito.when(kbDao.assertPackageExistence(packageName, version)).thenReturn(true);
+        KnowledgeBaseConnector.rcgBaseUrl = "http://lima.ewi.tudelft.nl/";
+        var result = service.getERCGLink(packageName, version, null, null);
+        assertNotNull(result);
+
+        Mockito.verify(kbDao, Mockito.times(1)).assertPackageExistence(packageName, version);
+
+    }
+
+    @Test
+    void getERCGLinkIngestionTest() throws IOException {
+        var packageName = "junit:junit";
+        var version = "4.12";
+        Mockito.when(kbDao.assertPackageExistence(packageName, version)).thenReturn(false);
+        var result = service.getERCGLink(packageName, version, null, null);
+        assertEquals(HttpStatus.CREATED, result.getStatusCode());
+
+        Mockito.verify(kbDao, Mockito.times(1)).assertPackageExistence(packageName, version);
     }
 }
